requirements.txt
openai
django-restframework
python-dotenv
requests